Refactor math/struct, add phi.backend
Add internal package phi.backend containing the Backend API and low-level helper functions shared among backends.
Move tensorop to phi.backend but keep reference in phi.struct for compatibility.

This refactoring will allow structs to use low-level math functions like shape/staticshape without cyclic dependencies.

Update affected documentation diagrams.
