+++ phi/tf/tf_backend.py
@@ -7,8 +7,8 @@ import six
 import tensorflow as tf
 from packaging import version
 
-from phi.math.base_backend import Backend
-from phi.struct.tensorop import expand, collapsed_gather_nd
+from phi.backend.backend import Backend
+from phi.backend.tensorop import expand, collapsed_gather_nd
 
 if tf.__version__[0] == '2':
     logging.info('Adjusting for tensorflow 2.0')
